+++ NEWS.md
@@ -1,6 +1,7 @@
 # cdom 0.1.1 (Unreleased)
 
-- `fit_exponential()` will now return `NULL` if the model do not converge. This make possible to use the function with the `apply` familly.
+- `cdom_fit_exponential()` has been renamed to `cdom_exponential()`.
+- Now using `purrr:safely()` to safely fit non-linear model.
  
 # cdom 0.1.0